Split tracking components
The only changes are temporary and caused by circular dependencies.
